orchestrator/requirements.txt
# LLM Orchestration
langgraph==0.2.28
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-google-genai==2.0.0
langchain-openai==0.2.1

# Tools & Sandbox
e2b==1.0.2
mcp==1.1.0

# RAG
llama-index==0.11.9
llama-index-vector-stores-postgres==0.3.0
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
aiofiles==24.1.0

# Observability
arize-phoenix==4.29.0
langsmith==0.1.129

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0
ruff==0.6.9
